backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
pydantic-settings>=2.0
aiofiles>=23.0
httpx>=0.27
greenlet>=3.0
aiosqlite>=0.22
pytest>=8.0
pytest-asyncio>=0.24
